feat: Redesign application layer
