+++ src/main/kotlin/dev/minuk/opampcommander/application/usecases/HandleAgentToServerUsecase.kt
@@ -1,0 +1,7 @@
+package dev.minuk.opampcommander.application.usecases
+
+import opamp.proto.Opamp.AgentToServer
+
+interface HandleAgentToServerUsecase {
+    suspend fun handleAgentToServer(agentToServer: AgentToServer)
+}
