+++ src/main/kotlin/dev/minuk/opampcommander/domain/models/agent/CustomMessage.kt
@@ -1,0 +1,7 @@
+package dev.minuk.opampcommander.domain.models.agent
+
+data class CustomMessage(
+    val capability: String,
+    val type: String,
+    val data: ByteArray,
+)
